Use fingerprinted file name for microfrontend entrypoint bundle
Enables us to add a caching header to the response
that fetches the entrypoint bundle of a microfrontend like.

Since the file name of the entrypoint bundle contains a md5 hash
based on the content of the file, we know that if a bundle with the same name
has been fetched by the browser, it is unchanged.

E.g. caching for one year:
Cache-Control: max-age=31536000

The main.js file name we used before did not allow for caching,
while still ensuring that all customers always get the latest micro frontend code.